chore: create abstraction to unify commands
