+++ scripts/tools/StuckTaskSummary.php
@@ -22,41 +22,10 @@ declare(strict_types=1);
 
 namespace oat\taoTaskQueue\scripts\tools;
 
-use oat\oatbox\extension\script\ScriptAction;
 use oat\oatbox\reporting\Report;
-use oat\taoTaskQueue\model\Repository\StuckTaskQuery;
-use oat\taoTaskQueue\model\Repository\StuckTaskRepository;
 
-class StuckTaskSummary extends ScriptAction
+class StuckTaskSummary extends AbstractStuckTask
 {
-    protected function provideOptions(): array
-    {
-        return [
-            'whitelist' => [
-                'prefix' => 'w',
-                'longPrefix' => 'whitelist',
-                'cast' => 'string',
-                'required' => true,
-                'description' => 'The whitelist of task_name that can be restarted.'
-            ],
-            'queue' => [
-                'prefix' => 'q',
-                'longPrefix' => 'queue',
-                'cast' => 'string',
-                'required' => true,
-                'description' => 'The queue to consider in the scope.'
-            ],
-            'age' => [
-                'prefix' => 'w',
-                'longPrefix' => 'age',
-                'cast' => 'int',
-                'required' => false,
-                'default' => 300,
-                'description' => 'Age in seconds of a task log.'
-            ]
-        ];
-    }
-
     protected function provideDescription(): string
     {
         return 'Summarize stuck tasks in the queue';
@@ -64,13 +33,8 @@ class StuckTaskSummary extends ScriptAction
 
     protected function run(): Report
     {
-        $query = new StuckTaskQuery(
-            $this->getOption('queue'),
-            explode(',', (string)$this->getOption('whitelist')),
-            $this->getOption('age')
-        );
-
-        $stuckTasks = $this->getStuckTasksRepository()->findAll($query);
+        $query = $this->getQuery();
+        $stuckTasks = $this->findStuckTasks();
 
         $report = Report::createSuccess(
             sprintf(
@@ -103,10 +67,5 @@ class StuckTaskSummary extends ScriptAction
         }
 
         return $report;
-    }
-
-    private function getStuckTasksRepository(): StuckTaskRepository
-    {
-        return $this->getServiceLocator()->get(StuckTaskRepository::class);
     }
 }
